refactor: add ApizrPackages enumeration for package management
